Search algorithm working fully, bar unknown bugs.

+++ app/src/main/java/android/example/wordlehelper/myMethods.kt
@@ -69,20 +69,24 @@ class myMethods {
         for (word in wordList) {
             word.lowercase().forEachIndexed() { index, letter ->
                 input.forEachIndexed(){inputIndex,inputLetter ->
-                    // Si la letra que toca corresponde a una verde en el input
-                    val comparacion = input[inputIndex].color
-                    // MAPAS: input.get busca el color (value) de LETTER (key) en mi mapa de input
-                    // Si no existe LETTER en mi input, devuelve null
-                    if (comparacion == android.example.wordlehelper.input.letterColor.GREEN && index == inputIndex && input[inputIndex].letter == letter){
-                        updatedList = deleteRestOfWords(updatedList,letter,index)
+                    if (!word.contains(inputLetter.toString())) updatedList.remove(word)
+                    else {
+
+                        // Si la letra que toca corresponde a una verde en el input
+                        val comparacion = input[inputIndex].color
+                        // MAPAS: input.get busca el color (value) de LETTER (key) en mi mapa de input
+                        // Si no existe LETTER en mi input, devuelve null
+                        if (comparacion == android.example.wordlehelper.input.letterColor.GREEN && index == inputIndex && input[inputIndex].letter == letter) {
+                            updatedList = deleteRestOfWordsGreen(updatedList, letter, index)
+                        }
+                        // Si la palabra contiene amarillas
+                        else if (comparacion == android.example.wordlehelper.input.letterColor.YELLOW && input[inputIndex].letter == letter) {
+                            updatedList = deleteRestOfWordsYellow(updatedList, letter, index)
+                        }
+                        // Si la palabra contiene negras
+                        else if (comparacion == android.example.wordlehelper.input.letterColor.BLACK && input[inputIndex].letter == letter)
+                            updatedList.remove(word)
                     }
-                    // Si la palabra contiene amarillas
-                    else if(comparacion == android.example.wordlehelper.input.letterColor.YELLOW && input[inputIndex].letter == letter){
-                        updatedList = deleteRestOfWordsYellow(updatedList,letter,index)
-                    }
-                    // Si la palabra contiene negras
-                    else if(comparacion == android.example.wordlehelper.input.letterColor.BLACK && input[inputIndex].letter == letter)
-                        updatedList.remove(word)
                 }
 
 
@@ -93,16 +97,22 @@ class myMethods {
     }
 
     /** This method removes from the list every word that doesn't fit the Green user inputs*/
-    fun deleteRestOfWords(
+    fun deleteRestOfWordsGreen(
         wordList: MutableList<String>,
         inputLetter: Char,
         posicion: Int
     ): MutableList<String> {
         val updatedList = wordList.toMutableList()
         for (word in wordList) {
-            word.lowercase().forEachIndexed() { index, letter ->
-                if (letter != inputLetter && posicion == index)
-                    updatedList.remove(word)
+            if (!word.contains(inputLetter.toString())) {
+                updatedList.remove(word)
+                println(word)
+            }
+            else {
+                word.lowercase().forEachIndexed() { index, letter ->
+                    if (letter != inputLetter && posicion == index)
+                        updatedList.remove(word)
+                }
             }
         }
         return updatedList.sorted().toMutableList()
@@ -116,14 +126,16 @@ class myMethods {
     ): MutableList<String> {
         val updatedList = wordList.toMutableList()
         for (word in wordList) {
-            var letterIsInWord = false
+            if(!word.contains(inputLetter.toString())) updatedList.remove(word)
+
+           /**var letterIsInWord = false
             word.lowercase().forEachIndexed() { index, letter ->
                 if (letter == inputLetter) letterIsInWord = true
             }
             if (!letterIsInWord) {
                 updatedList.remove(word)
                 letterIsInWord = false
-            }
+            }*/
         }
         return updatedList.sorted().toMutableList()
     }
